Merge '2.14.0-188.0.dev' into beta

+++ pkg/vm_snapshot_analysis/test/utils.dart
@@ -24,7 +24,7 @@ class AotSnapshot {
   final String outputBinary;
   final String sizesJson;
 
-  AotSnapshot({this.outputBinary, this.sizesJson});
+  AotSnapshot({required this.outputBinary, required this.sizesJson});
 }
 
 Future withFlag(
@@ -33,7 +33,7 @@ Future withFlag(
 }
 
 Future withFlagImpl(
-    Map<String, String> source, String flag, Future Function(AotSnapshot) f) {
+    Map<String, String> source, String? flag, Future Function(AotSnapshot) f) {
   return withTempDir((dir) async {
     final snapshot = AotSnapshot(
       outputBinary: path.join(dir, 'output.exe'),
@@ -60,17 +60,21 @@ void main(List<String> args) => input.main(args);
       if (flag != null) '$flag=${snapshot.sizesJson}',
     ];
 
-    // Compile input.dart to native and output instruction sizes.
-    final result = await Process.run(dart2native, [
+    final args = [
       '-o',
       snapshot.outputBinary,
       '--packages=$packages',
       '--extra-gen-snapshot-options=${extraGenSnapshotOptions.join(',')}',
       mainDart,
-    ]);
+    ];
+
+    // Compile input.dart to native and output instruction sizes.
+    final result = await Process.run(dart2native, args);
 
     expect(result.exitCode, equals(0), reason: '''
-Compilation completed successfully.
+Compilation completed with exit code ${result.exitCode}.
+
+Command line: $dart2native ${args.join(' ')}
 
 stdout: ${result.stdout}
 stderr: ${result.stderr}
@@ -86,20 +90,25 @@ stderr: ${result.stderr}
   });
 }
 
+late final shouldKeepTemporaryDirectories =
+    Platform.environment['KEEP_TEMPORARY_DIRECTORIES']?.isNotEmpty == true;
+
 Future withTempDir(Future Function(String dir) f) async {
   final tempDir =
       Directory.systemTemp.createTempSync('instruction-sizes-test-');
   try {
     await f(tempDir.path);
   } finally {
-    tempDir.deleteSync(recursive: true);
+    if (shouldKeepTemporaryDirectories) {
+      tempDir.deleteSync(recursive: true);
+    }
   }
 }
 
 Future<Object> loadJson(File input) async {
-  return await input
+  return (await input
       .openRead()
       .transform(utf8.decoder)
       .transform(json.decoder)
-      .first;
+      .first)!;
 }
